refactor: rename message files
