+++ cli/src/shared/messages/generate.message.ts
