+++ cli/src/shared/messages/index.ts
@@ -1,9 +1,10 @@
-export { buildMessage } from './build.messages';
-export * from './common.messages';
-export { addMessage } from './add.messages';
-export { generateMessage } from './generate.messages';
-export { independencyMessage } from './independency.messages';
-export { installMessage } from './install.messages';
-export { lintMessage } from './lint.messages';
-export { serveMessage } from './serve.messages';
-export { workspacesMessage } from './workspaces.messages';
+export * from './common.message';
+export { addMessage } from './add.message';
+export { buildMessage } from './build.message';
+export { generateMessage } from './generate.message';
+export { independencyMessage } from './independency.message';
+export { installMessage } from './install.message';
+export { lintMessage } from './lint.message';
+export { serveMessage } from './serve.message';
+export { submodulesMessage } from './submodules.message';
+export { workspacesMessage } from './workspaces.message';
